refactor: move test outside testclasses
