+++ src/test/java/spoon/test/position/PositionTest.java
@@ -59,7 +59,7 @@ import spoon.test.position.testclasses.PositionParameterTypeWithReference;
 import spoon.test.position.testclasses.PositionTry;
 import spoon.test.position.testclasses.SomeEnum;
 import spoon.test.position.testclasses.TypeParameter;
-import spoon.test.query_function.testclasses.VariableReferencesModelTest;
+import spoon.test.query_function.VariableReferencesModelTest;
 import spoon.testing.utils.ModelUtils;
 
 import java.io.File;
